% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bmi.ais}
\alias{bmi.ais}
\title{Effect of Body Mass Index (BMI) on Acute Ischemic Stroke (AIS)}
\format{
A \code{data.frame} with 1880 rows and 29 variables.
}
\usage{
data(bmi.ais)
}
\description{
This dataset is created from three genome-wide association studies:
\enumerate{
\item A 2017 GWAS of BMI by Akiyama et al.
\item The UK BioBank GWAS of BMI (2nd round release by the Neale lab).
\item A 2018 GWAS of AIS by Malik et al.
}
To obtain this dataset, the Akiyama study is used for SNP selection (column \code{pval.selection}). The UK BioBank dataset estimates the SNPs' effect on BMI and the Malik dataset estimates the SNPs' effect on AIS.
}
\keyword{datasets}
