% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bmi.bmi}
\alias{bmi.bmi}
\title{"Effect" of Body Mass Index (BMI) on Body Mass Index (BMI)}
\format{A data.frame.}
\usage{
data(bmi.bmi)
}
\description{
Summary data obtained by combining three genome-wide association studies:
\enumerate{
\item{BMI-GIANT:}{ BMI in the Genetic Investigation of ANthropometric Traits (GIANT) consortium (sample size: 339224).}
\item{BMI-UKBB-1:}{ BMI in a half of the United Kingdom BioBank (UKBB) data (sample size: 234070)}
\item{SBP-UKBB-2:}{ BMI in the other half of the UKBB data (sample size: 234070)}
}
}
\details{
The BMI-GIANT dataset is used for SNP selection (column \code{pval.selection}). The BMI-UKBB-1 dataset estimates the SNPs' effects on BMI (columns \code{beta.exposure} and \code{se.exposure}) and the BMI-UKBB-2 dataset provides independent estimates of the same effects (columns \code{beta.outcome} and \code{se.outcome}).
}
\keyword{datasets}
