% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bmi.cad}
\alias{bmi.cad}
\title{Effect of Body Mass Index (BMI) on Coronary Artery Disease (CAD)}
\format{
A \code{data.frame} with 1119 rows and 42 variables.
}
\usage{
data(bmi.cad)
}
\description{
This dataset is created from three genome-wide association studies:
\enumerate{
\item A 2017 GWAS of BMI by Akiyama et al.
\item The UK BioBank GWAS of BMI (2nd round release by the Neale lab).
\item The CARDIoGRAMplusC4D with 1000 Genome Project imputation GWAS of CAD.
}
To obtain this dataset, the Akiyama study is used for SNP selection (column \code{pval.selection}). The UK BioBank dataset estimates the SNPs' effect on BMI and the CARDIoGRAMplusC4D estimates the SNPs' effect on CAD.
}
\keyword{datasets}
