% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bmi.sbp}
\alias{bmi.sbp}
\title{Effect of Body Mass Index (BMI) on Systolic Blood Pressure (SBP)}
\format{A data.frame.}
\usage{
data(bmi.sbp)
}
\description{
Summary data obtained by combining three genome-wide association studies:
\enumerate{
\item{BMI-FEM:}{ BMI in females by the Genetic Investigation of ANthropometric Traits (GIANT) consortium (sample size: 171977).}
\item{BMI-MAL:}{ BMI in males in the same study by the GIANT consortium (sam- ple size: 152893)}
\item{SBP-UKBB:}{ SBP using the United Kingdom BioBank (UKBB) data (sample size: 317754)}
}
}
\details{
The BMI-FEM dataset is used for SNP selection (column \code{pval.selection}). The BMI-MAL dataset estimates the SNPs' effect on BMI and the SBP-UKBB dataset estimates the SNPs' on SBP.
}
\keyword{datasets}
