% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cad.bmi}
\alias{cad.bmi}
\title{Effect of Coronary Artery Disease (CAD) on Body Mass Index (BMI) (To study reverse causality)}
\format{A \code{data.frame} with 1625 rows and 34 variables.}
\usage{
data(cad.bmi)
}
\description{
This dataset is created from three genome-wide association studies:
\enumerate{
\item The C4D GWAS of CAD.
\item The CARDIoGRAM GWAS of CAD.
\item The UK BioBank GWAS of BMI (2nd round release by the Neale lab).
}
To obtain this dataset, the C4D study is used for SNP selection (column \code{pval.selection}). The CARDIoGRAM dataset estimates the SNPs' effect on CAD and the UK BioBank dataset estimates the SNPs' on BMI.
}
\keyword{datasets}
