% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cad.cad}
\alias{cad.cad}
\title{Effect of Coronary Artery Disease (BMI) on Coronary Artery Disease (CAD)}
\format{A \code{data.frame} with 1650 rows and 39 variables.}
\usage{
data(cad.cad)
}
\description{
This dataset is created from three genome-wide association studies:
\enumerate{
\item The UK BioBank GWAS of heart attack (2nd round release by the Neale lab).
\item The C4D GWAS of CAD.
\item The CARDIoGRAM GWAS of CAD.
}
This dataset serves the purpose of a validation study. Since both the "exposure" and the "outcome" are CAD, the joint normal model of the GWAS summary statistics is expected to hold with no pleiotropy and "causal effect" equal to 1.
}
\keyword{datasets}
