% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crp.cad}
\alias{crp.cad}
\title{Effect of C-Reactive Protein on Coronary Artery Disease (CAD)}
\format{A \code{data.frame} with 1575 rows and 30 variables.}
\usage{
data(crp.cad)
}
\description{
This dataset is created from three genome-wide association studies:
\enumerate{
\item Prins et al.\ (2017) GWAS of CRP.
\item Dehghan et al.\ (2011) GWAS of CRP
\item The CARDIoGRAMplusC4D GWAS of CAD.
}
To obtain this dataset, the Prins study is used for SNP selection (column \code{pval.selection}). The Dehghan dataset estimates the SNPs' effect on CRP and the CARDIoGRAMplusC4D dataset estimates the SNPs' on CAD.
}
\keyword{datasets}
