% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crp.mi}
\alias{crp.mi}
\title{Effect of C-Reactive Protein on heart attack}
\format{A \code{data.frame} with 1581 rows and 31 variables.}
\usage{
data(crp.mi)
}
\description{
This dataset is created from three genome-wide association studies:
\enumerate{
\item Prins et al.\ (2017) GWAS of CRP.
\item Dehghan et al.\ (2011) GWAS of CRP
\item The UK BioBank GWAS of self reported heart attack (2nd round release by the Neale lab).
}
To obtain this dataset, the Prins study is used for SNP selection (column \code{pval.selection}). The Dehghan dataset estimates the SNPs' effect on CRP and the UK BioBank dataset estimates the SNPs' on heart attack.
}
\keyword{datasets}
