% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkage.R
\name{fit.mixture.model}
\alias{fit.mixture.model}
\title{Fit a Gaussian mixture deconvolution model}
\usage{
fit.mixture.model(z, n = 2, ntry = 10, force.mu.zero = TRUE,
  diagnosis = FALSE)
}
\arguments{
\item{z}{A vector of z-scores.}

\item{n}{Number of mixture components.}

\item{ntry}{Number of random initializations.}

\item{force.mu.zero}{Should the means be forced to zero?}

\item{diagnosis}{Logical indicator for showing diagnostic plots.}
}
\value{
A list of \code{p} (mixture proportion), \code{mu} (mean), \code{sigma} (standard deviation).
}
\description{
Fit a Gaussian mixture deconvolution model
}
\details{
This function assumes that \eqn{z} is distributed as \eqn{N(\gamma, 1)} and \eqn{\gamma} follows a Gaussian mixture model. It fits this deconvolution model by maximum likelihood and outputs the estimated mixture distribution.
}

