% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{ivw.modified}
\alias{ivw.modified}
\title{Modified weights IVW}
\usage{
ivw.modified(b_exp, b_out, se_exp, se_out)
}
\arguments{
\item{b_exp}{A vector of SNP effects on the exposure variable, usually obtained from a GWAS.}

\item{b_out}{A vector of SNP effects on the outcome variable, usually obtained from a GWAS.}

\item{se_exp}{A vector of standard errors of \code{b_exp}.}

\item{se_out}{A vector of standard errors of \code{b_out}.}
}
\value{
A list
\describe{
\item{beta.hat}{Estimated causal effect}
\item{beta.se}{Standard error of \code{beta.hat}}
}
}
\description{
This function implements the modified 2nd order weighted procedure.
}
\references{
Bowden, Jack, M. Fabiola Del Greco, Cosetta Minelli, Debbie Lawlor, Nuala Sheehan, John Thompson, and George Davey Smith. "Improving the accuracy of two-sample summary data Mendelian randomization: moving beyond the NOME assumption." bioRxiv (2017): 159442.
}

