% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lipid.cad}
\alias{lipid.cad}
\title{Effect of blood lipids (LDL cholesterol, HDL cholesterol, Triglycerides) on Cardiovascular disease risk}
\format{A \code{data.frame} with 10168 rows and 24 variables.}
\usage{
data(lipid.cad)
}
\description{
This dataset is created from four genome-wide association studies:
\enumerate{
\item A 2010 GWAS of blood lipids (Teslovich et al.\, 2010), named "teslovich_2010" in the dataset.
\item The MetaboChip (MC) data in a 2013 GWAS of blood lipids (Willer et al.\, 2013), named "mc_2013" in the dataset.
\item The CARDIoGRAMplusC4D meta-analysis of coronary artery disease (CARDIoGRAMplusC4D Consortium, 20135, named "cardiogramplusc4d_1000genome" in the dataset.
\item The UK BioBank GWAS of self reported heart attach (interim release by the Neale lab), named "ukbb_6150_round2" in the dataset.
}
}
\details{
\code{lipid.cad} contains in total 24 sub-datasets, each is suitable for a Mendelian randomization study. To obtain a sub-dataset, you must decide on
\describe{
\item{lipid}{Which lipid trait to consider? Either \code{ldl}, \code{hdl}, or \code{tg}.}
\item{gwas.selection}{Which GWAS is used for selection? Either \code{teslovich_2010} or \code{mc_2013}.}
\item{gwas.exposure}{Which GWAS is used for exposure? Either \code{teslovich_2010} or \code{mc_2013} and must be different from \code{gwas.selection}.}
\item{gwas.outcome}{Which GWAS is used for outcome? Either \code{cardiogramplusc4d} or \code{ukbb_self_report_heart}.}
\item{restrict}{Should we use SNPs that are not associated with the other lipids? For example, if we are studying the effect of HDL cholesterol (so \code{lipid} is "hdl") and \code{restrict} is TRUE, then the SNPs are not associated with LDL cholesterol and triglycerides (p-value > 0.01 in the \code{gwas.selection} data).}
}
}
\keyword{datasets}
