% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{modal.plot}
\alias{modal.plot}
\title{Modal plot to detect heterogeneity}
\usage{
modal.plot(
  b_exp = NULL,
  b_out = NULL,
  se_exp = NULL,
  se_out = NULL,
  data = NULL,
  k = 1.5,
  weight.option = c("MLE", "shrinkage"),
  beta.range = NULL
)
}
\arguments{
\item{b_exp}{A vector of SNP effects on the exposure variable, usually obtained from a GWAS.}

\item{b_out}{A vector of SNP effects on the outcome variable, usually obtained from a GWAS.}

\item{se_exp}{A vector of standard errors of \code{b_exp}.}

\item{se_out}{A vector of standard errors of \code{b_out}.}

\item{data}{Alternatively, dataset can be passed by the argument \code{data}, which must be a data frame with columns \code{beta.exposure}, \code{beta.outcome}, \code{se.exposure}, \code{se.outcome}.}

\item{k}{Locality of the robust likelihood (smaller \code{k} has more sensitivity for mode detection)}

\item{weight.option}{Character. Choice of \code{"MLE"} or \code{"shrinkage"}.}

\item{beta.range}{range of beta in the plot}
}
\description{
Modal plot to detect heterogeneity
}
\examples{
data(lipid.cad)
data <- subset(lipid.cad, lipid == "hdl" & restrict &
gwas.selection == "teslovich_2010" &
gwas.outcome == "cardiogramplusc4d_1000genome" &
pval.selection < 1e-5)

modal.plot(data$beta.exposure, data$beta.outcome, data$se.exposure, data$se.outcome, k = 1)

data <- subset(lipid.cad, lipid == "ldl" & restrict &
gwas.selection == "teslovich_2010" &
gwas.outcome == "cardiogramplusc4d_1000genome" &
pval.selection < 1e-5)

modal.plot(data$beta.exposure, data$beta.outcome, data$se.exposure, data$se.outcome)

}
