% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.raps.R
\docType{package}
\name{mr.raps-package}
\alias{mr.raps-package}
\title{Two Sample Mendelian Randomization using Robust Adjusted Profile Score}
\description{
Mendelian randomization is a method of identifying and estimating a confounded causal effect using genetic instrumental variables. This packages implements methods for two sasmple Mendelian randomization with summary statistics by using Robust Adjusted Profile Score (RAPS).
}
