% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkage.R
\name{mr.raps.publish}
\alias{mr.raps.publish}
\title{Generate diagnostic plots for publishing}
\usage{
mr.raps.publish(data, weight.option = c("both", "MLE", "shrinkage"))
}
\description{
Generate diagnostic plots for publishing
}
\keyword{internal}
