% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkage.R
\name{mr.raps}
\alias{mr.raps}
\title{Recommended \code{mr.raps} procedure}
\usage{
mr.raps(
  data,
  diagnostics = TRUE,
  over.dispersion = TRUE,
  loss.function = "huber",
  ...
)
}
\arguments{
\item{data}{A data frame (see Details)}

\item{diagnostics}{Logical indicator for showing diagnostic plots.}

\item{over.dispersion}{Should the model consider overdispersion (systematic pleiotropy)? Default is FALSE.}

\item{loss.function}{Either the squared error loss (\code{l2}) or robust loss functions/scores (\code{huber} or \code{tukey}).}

\item{...}{Additional parameters to be passed to \code{mr.raps.shrinkage} (default is \code{shrinkage=FALSE}).}
}
\description{
Recommended \code{mr.raps} procedure
}
\details{
This function calls \code{mr.raps.shrinkage} with \code{overdispersion = TRUE}, \code{loss.function = "huber"}. The input data frame should contain the following variables:
\enumerate{
\item beta.exposure
\item beta.outcome
\item se.exposure
\item se.outcome
}
}
\examples{

mr.raps(bmi.sbp)

\donttest{
require(mr.raps)
data(lipid.cad)
data <- subset(lipid.cad, lipid == "hdl" & restrict &
gwas.selection == "teslovich_2010" &
gwas.outcome == "cardiogramplusc4d_1000genome")
mr.raps(data)
}

}
\seealso{
mr.raps.shrinkage
}
