% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkage.R
\name{mr.raps.recommend}
\alias{mr.raps.recommend}
\title{Recommended \code{mr.raps} procedure}
\usage{
mr.raps.recommend(data, verbose = TRUE)
}
\arguments{
\item{data}{A data frame (see Details)}

\item{verbose}{Print the result}
}
\description{
Recommended \code{mr.raps} procedure
}
\details{
The data frame should contain the following variables:
\enumerate{
\item beta.exposure
\item beta.outcome
\item se.exposure
\item se.outcome
}
}
