% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mr.raps.scatterplot}
\alias{mr.raps.scatterplot}
\title{Scatter plot with annotation}
\usage{
mr.raps.scatterplot(data, annotate = TRUE, annotate.genes = NULL,
  rank.method = c("pval.both", "pval.selection", "pval.exposure"),
<<<<<<< HEAD
  num.snps = 10, fit = mr.raps(data, FALSE), alpha = 0.8)
=======
  num.snps = 10, fit = mr.raps(data, FALSE))
>>>>>>> master
}
\arguments{
\item{data}{A data frame (see \code{\link{mr.raps}}).}

\item{annotate}{Annotating the points? (rsid, chromosome, position)}

\item{annotate.genes}{Further annotation of closest genes? See example.}

\item{rank.method}{How to select strongest SNPs for plot?}

\item{num.snps}{How many SNPs are shown?}

\item{fit}{A \code{mr.raps} fit.}
}
\description{
Scatter plot with annotation
}
\examples{
data(bmi.sbp)
mr.raps.scatterplot(bmi.sbp)

\donttest{
require(bumphunter)
require(TxDb.Hsapiens.UCSC.hg38.knownGene)
genes <- annotateTranscripts(TxDb.Hsapiens.UCSC.hg38.knownGene)
mr.raps.scatterplot(bmi.sbp, annotate.genes = genes)
}

}
