% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkage.R
\name{mr.raps.shrinkage}
\alias{mr.raps.shrinkage}
\title{Main function (for RAPS with shrinkage)}
\usage{
mr.raps.shrinkage(b_exp, b_out, se_exp, se_out, over.dispersion = FALSE,
  loss.function = c("l2", "huber", "tukey"), k = switch(loss.function[1], l2
  = 2, huber = 1.345, tukey = 4.685), shrinkage = TRUE, prior.param = NULL,
  diagnosis = FALSE, se.method = c("sandwich", "bootstrap"),
  num.init = 10)
}
\arguments{
\item{b_exp}{A vector of SNP effects on the exposure variable, usually obtained from a GWAS.}

\item{b_out}{A vector of SNP effects on the outcome variable, usually obtained from a GWAS.}

\item{se_exp}{A vector of standard errors of \code{b_exp}.}

\item{se_out}{A vector of standard errors of \code{b_out}.}

\item{over.dispersion}{Should the model consider overdispersion (systematic pleiotropy)? Default is FALSE.}

\item{loss.function}{Either the squared error loss (\code{l2}) or robust loss functions/scores (\code{huber} or \code{tukey}).}

\item{k}{Threshold parameter in the Huber and Tukey loss functions.}

\item{shrinkage}{If shrinkage (empirical partially Bayes) should be used. Shrinkage does not affect the unbiasedness of the estimating equations and generally will increase the estimation accuracy.}

\item{prior.param}{Parameters of the Gaussian spike-and-slab prior}

\item{diagnosis}{Should the function returns diagnostic plots and results? Default is FALSE}

\item{se.method}{How should the standard error be estimated? Either by sandwich variance formula (default and recommended) or the bootstrap.}

\item{num.init}{Number of initializations.}
}
\description{
Main function (for RAPS with shrinkage)
}
\details{
\code{mr.raps.shrinkage} is the main function for RAPS in conjunction with empirical partially Bayes. It is more general than the first generation \code{mr.raps} function and should be preferred in practice. With the option \code{shrinkage = TRUE}, it essentially reduces to \code{mr.raps}. In that case, the main difference is that the standard errors in \code{mr.raps.shrinkage} are computed based on observed information (and also an empirical estimate of the variance of the score function). This is preferred over using the plugged-in Fisher information in \code{mr.raps}. See Efron and Hinkley (1978) referenced below.

Because the estimating equations are highly non-linear, it is possible that there are multiple roots. To overcome this issue, we use multiple initializations (controlled by \code{num.init}) around the \code{mr.raps} point estimate. A warning is given if there seems to be another finite root, and no solution is returned if there are two roots close to the initialization.
}
\examples{

require(mr.raps)
data(lipid.cad)
data <- subset(lipid.cad, lipid == "hdl" & restrict & gwas.selection == "teslovich_2010" & gwas.outcome == "cardiogramplusc4d")
z <- data$beta.exposure / data$se.exposure
prior.param <- fit.mixture.model(z)

mr.raps.shrinkage(data$beta.exposure, data$beta.outcome, data$se.exposure, data$se.outcome, TRUE, "huber", shrinkage = FALSE)
mr.raps.shrinkage(data$beta.exposure, data$beta.outcome, data$se.exposure, data$se.outcome, TRUE, "huber", shrinkage = TRUE,
prior.param = prior.param)

## Diagnostic plots
out <- mr.raps.shrinkage(data$beta.exposure, data$beta.outcome, data$se.exposure, data$se.outcome, TRUE, "huber",
shrinkage = TRUE, prior.param = prior.param, diagnosis = TRUE)

}
\references{
Qingyuan Zhao, Q., Chen, Y., Wang, J., and Small, D. S. (2018). A genome-wide design and an empirical partially Bayes approach to increase the power of Mendelian randomization, with application to the effect of blood lipids on cardiovascular disease. <arXiv:1804.07371>.
Efron, B. and Hinkley, D. V. (1978). Assessing the accuracy of the maximum likelihood estimator: Observed versus expected Fisher information. Biometrika, 65(3), 457--483.
}
