% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkage.R
\name{posterior.mean}
\alias{posterior.mean}
\title{Compute the posterior mean under spike-and-slab Gaussian prior}
\usage{
posterior.mean(z, sigma, p, mu, sigma.prior, deriv = 0)
}
\arguments{
\item{z}{a vector of z-scores}

\item{sigma}{a vector of standard deviations of \code{z} (if \code{sigma} is a single number, it is expanded to a vector)}

\item{p}{prior: mixture proportion}

\item{mu}{prior: mean}

\item{sigma.prior}{prior: standard deviation}

\item{deriv}{compute the posterior mean (\code{deriv = 0}) or its derivative (\code{deriv = 1})}
}
\value{
a vector
}
\description{
Compute the posterior mean under spike-and-slab Gaussian prior
}
\details{
Similar to \code{fit.mixture.model}, this function assumes that \eqn{z} is distributed as \eqn{N(\gamma, 1)} and \eqn{\gamma} follows a Gaussian mixture model. The function computes the posterior mean \eqn{E[\gamma|z]}.
}
\examples{

require(mr.raps)
data(lipid.cad)
data <- subset(lipid.cad, lipid == "hdl" & restrict &
gwas.selection == "teslovich_2010" &
gwas.outcome == "cardiogramplusc4d_1000genome")
z <- data$beta.exposure / data$se.exposure
prior.param <- fit.mixture.model(z)

z.seq <- seq(-5, 5, 0.1)
gamma.hat <- posterior.mean(z.seq, 1, prior.param$p, prior.param$mu, prior.param$sigma)
gamma.hat.deriv <- posterior.mean(z.seq, 1, prior.param$p,
prior.param$mu, prior.param$sigma, deriv = 1)
par(mfrow = c(1, 2))
plot(z.seq, gamma.hat, type = "l")
plot(z.seq, gamma.hat.deriv, type = "l")

}
