% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R, R/mr_fun_general.R
\name{rho.l2}
\alias{rho.l2}
\alias{rho.l2}
\title{Squared error loss and its derivatives}
\usage{
rho.l2(r, deriv = 0)

rho.l2(r, deriv = 0)
}
\description{
Squared error loss and its derivatives

L2 loss function and its derivatives
}
\keyword{internal}
