% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{modal.plot}
\alias{modal.plot}
\title{Modal plot to detect heterogeneity}
\usage{
modal.plot(b_exp = NULL, b_out = NULL, se_exp = NULL, se_out = NULL,
  data = NULL, k = 1.5, weight.option = c("MLE", "shrinkage"))
}
\arguments{
\item{data}{Alternatively, dataset can be passed by the argument \code{data}, which must be a data frame with columns \code{beta.exposure}, \code{beta.outcome}, \code{se.exposure}, \code{se.outcome}.}

\item{k}{Locality of the robust likelihood (smaller \code{k} has more sensitivity for mode detection)}
}
\description{
Modal plot to detect heterogeneity
}
\examples{
data(lipid.cad)
data <- subset(lipid.cad, lipid == "hdl" & restrict &
gwas.selection == "teslovich_2010" &
gwas.outcome == "cardiogramplusc4d_1000genome" &
pval.selection < 1e-5)

modal.plot(data$beta.exposure, data$beta.outcome, data$se.exposure, data$se.outcome, k = 1)

data <- subset(lipid.cad, lipid == "ldl" & restrict &
gwas.selection == "teslovich_2010" &
gwas.outcome == "cardiogramplusc4d_1000genome" &
pval.selection < 1e-5)

modal.plot(data$beta.exposure, data$beta.outcome, data$se.exposure, data$se.outcome)

}
