% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkage.R
\name{mr.raps}
\alias{mr.raps}
\title{Recommended \code{mr.raps} procedure}
\usage{
mr.raps(data, diagnosis = TRUE)
}
\arguments{
\item{data}{A data frame (see Details)}
}
\description{
Recommended \code{mr.raps} procedure
}
\details{
This function calls \code{mr.raps.shrinkage} with \code{overdispersion = TRUE}, \code{loss.function = "huber"}, \code{shrinkage = TRUE}. The input data frame should contain the following variables:
\enumerate{
\item beta.exposure
\item beta.outcome
\item se.exposure
\item se.outcome
}
}
\examples{

require(mr.raps)
data(lipid.cad)
data <- subset(lipid.cad, lipid == "hdl" & restrict & gwas.selection == "teslovich_2010" & gwas.outcome == "cardiogramplusc4d")
mr.raps(data)

}
